% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{calc_d_balance}
\alias{calc_d_balance}
\title{Calculate D' Balance for runner in race}
\usage{
calc_d_balance(
  runner_data,
  race_data,
  runner_col_names = FALSE,
  race_col_names = FALSE
)
}
\arguments{
\item{runner_data}{a dataframe containing runners critical speed and D' prime values.}

\item{race_data}{a dataframe containing details of the race to simulate where each row represents a lap split. Needs lap speed & lap distance columns.}

\item{runner_col_names}{FALSE (default) or a character vector with the names to be replaced.
runner_col_names[[1]] is renamed to "athlete"
runner_col_names[[2]] is renamed to "critical_speed_meters_second"
runner_col_names[[3]] is renamed to  "d_prime_meters".
So if you table has "Athletes", "CS", and "DPrime" as names you should pass
runner_col_names =  c("Athletes", "CS", "DPrime") so the function knows which column is which.

 @param race_col_names FALSE (default) or a character vector with the names to be replaced.
race_col_names[[1]] is renamed to "lap_leader_speed_meters_second"
race_col_names[[2]] is renamed to "lap_distance_meters"}
}
\value{
a data frame in long format
}
\description{
Calculate D' Balance for runner in race
}
